requirements.txt
streamlit
openai
llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-pinecone
llama-index-llms-groq
scrapy